OD-11295 Round up average scores value (by default int/int division just throws away float part: 49/10=4)
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@54284 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ portal/src/main/java/ish/oncourse/portal/services/survey/GetAverageSurvey.java
@@ -63,16 +63,16 @@ public class GetAverageSurvey {
         }
         
         if (result.getVenueScore() > 0) {
-            result.setVenueScore((Math.round(result.getVenueScore() / venueScores)));
+            result.setVenueScore((int)(Math.round((double)result.getVenueScore() / venueScores)));
         }
         if (result.getCourseScore() > 0) {
-            result.setCourseScore((Math.round(result.getCourseScore() / courseScores)));
+            result.setCourseScore((int)(Math.round((double)result.getCourseScore() / courseScores)));
         }   
         if (result.getTutorScore() > 0) {
-            result.setTutorScore((Math.round(result.getTutorScore() / tutorScores)));
+            result.setTutorScore((int)(Math.round((double)result.getTutorScore() / tutorScores)));
         }   
         if (result.getNetPromoterScore() > 0) {
-            result.setNetPromoterScore((Math.round(result.getNetPromoterScore() / netPromoterScores)));
+            result.setNetPromoterScore((int)(Math.round((double)result.getNetPromoterScore() / netPromoterScores)));
         }
         
         return result;
